Split and rename bal-recovery
